agent/requirements.txt
# AWS Agentcore and Strands
strands-agents==1.20.0
bedrock-agentcore==1.1.1
aws-opentelemetry-distro==0.12.0
boto3==1.42.2
jinja2==3.1.6

# Geospatial Analysis
numpy==2.2.6
scipy==1.17.1
pystac==1.14.1
pystac_client==0.9.0
xarray==2025.6.1
stackstac==0.5.1
shapely==2.1.2
pyproj==3.7.0
rasterio==1.4.4
rio_tiler==8.0.5
rioxarray==0.18.2
httpx==0.28.1
pillow==12.1.0
geopandas==1.0.1

# Visualization
matplotlib==3.10.7
seaborn==0.13.2
